Rename nav directory to nav-bar to prepare for nav-panel
